{{ cookiecutter.repo_name }}/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
alembic=1.7.6=pypi_0
antlr4-python3-runtime=4.8=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h2bbff1b_0
astunparse=1.6.3=pypi_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bleach=4.1.0=pyhd3eb1b0_0
ca-certificates=2021.10.26=haa95532_4
cachetools=5.0.0=pypi_0
certifi=2021.10.8=pypi_0
cffi=1.15.0=py38h2bbff1b_1
charset-normalizer=2.0.12=pypi_0
click=8.0.3=pypi_0
cloudpickle=2.0.0=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
croniter=1.2.0=pypi_0
cycler=0.11.0=pypi_0
dask=2022.2.0=pypi_0
databricks-cli=0.16.4=pypi_0
debugpy=1.5.1=py38hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
distributed=2022.2.0=pypi_0
docker=5.0.3=pypi_0
entrypoints=0.4=pypi_0
et-xmlfile=1.1.0=pypi_0
flask=2.0.3=pypi_0
flatbuffers=2.0=pypi_0
fonttools=4.29.1=pypi_0
fsspec=2022.1.0=pypi_0
gast=0.5.3=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.26=pypi_0
google-auth=2.6.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
greenlet=1.1.2=pypi_0
grpcio=1.43.0=pypi_0
h5py=3.6.0=pypi_0
heapdict=1.0.1=pypi_0
hydra-core=1.1.1=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.1=pypi_0
importlib-resources=5.4.0=pypi_0
importlib_metadata=4.8.2=hd3eb1b0_0
ipykernel=6.4.1=py38haa95532_1
ipython=7.31.1=py38haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
itsdangerous=2.0.1=pypi_0
jedi=0.18.1=py38haa95532_1
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_core=4.9.1=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
keras=2.8.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
libclang=13.0.0=pypi_0
locket=0.2.1=pypi_0
mako=1.1.6=pypi_0
markdown=3.3.6=pypi_0
markupsafe=2.0.1=py38h2bbff1b_0
marshmallow=3.14.1=pypi_0
marshmallow-oneofschema=3.0.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py38he774522_1000
mlflow=1.23.1=pypi_0
msgpack=1.0.3=pypi_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.11=pyhd3eb1b0_0
nbconvert=6.1.0=py38haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.8=py38haa95532_0
numpy=1.22.2=pypi_0
oauthlib=3.2.0=pypi_0
omegaconf=2.1.1=pypi_0
openpyxl=3.0.9=pypi_0
openssl=1.1.1m=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
partd=1.2.0=pypi_0
pendulum=2.1.2=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=pypi_0
pip=21.2.2=py38haa95532_0
prefect=0.15.13=pypi_0
prometheus-flask-exporter=0.18.7=pypi_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydot=1.4.2=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.7=pypi_0
pyrsistent=0.18.0=py38h196d8e1_0
python=3.8.12=h6244533_0
python-box=5.4.1=pypi_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-graphviz=0.19.1=pypi_0
python-slugify=5.0.2=pypi_0
pytz=2021.3=pypi_0
pytzdata=2020.1=pypi_0
pywin32=227=pypi_0
pywinpty=2.0.2=py38h5da7b33_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py38hd77b12b_2
querystring-parser=1.2.4=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=60.9.2=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlalchemy=1.4.31=pypi_0
sqlite=3.37.2=h2bbff1b_0
sqlparse=0.4.2=pypi_0
tabulate=0.8.9=pypi_0
tblib=1.7.0=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-cpu=2.8.0=pypi_0
tensorflow-io-gcs-filesystem=0.24.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.1=py38haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
text-unidecode=1.3=pypi_0
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
threadpoolctl=3.1.0=pypi_0
toml=0.10.2=pypi_0
toolz=0.11.2=pypi_0
tornado=6.1=py38h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.0=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.8=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
waitress=2.0.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=1.2.3=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pypi_0
wincertstore=0.2=py38haa95532_2
winpty=0.4.3=4
wrapt=1.13.3=pypi_0
zict=2.0.0=pypi_0
zipp=3.7.0=pyhd3eb1b0_0
